requirements.txt
pandas==2.2.2 
sqlalchemy==2.0.31 
pydantic==2.8.2 
pydantic-settings==2.3.4 
python-dotenv==1.0.1 
psycopg2-binary==2.9.9